+++ python/paddle/fluid/tests/unittests/mkldnn/test_activation_mkldnn_op.py
@@ -79,90 +79,6 @@ class TestMKLDNNGeluDim2Approx(TestActivation):
         self.attrs = {"use_mkldnn": True, "approximate": True}
 
 
-#Use it as a base class for BF16 activation tests, just override necessary functions
-class TestMKLDNNSigmoidBF16Op(TestActivation):
-    @OpTestTool.skip_if_not_cpu_bf16()
-    def config(self):
-        self.op_type = "sigmoid"
-
-    def op_forward(self, x):
-        return 1 / (1 + np.exp(-x))
-
-    def op_grad(self, dout, x):
-        return dout * self.op_forward(x) * (1 - self.op_forward(x))
-
-    def set_attrs(self):
-        self.attrs = {"use_mkldnn": True}
-
-    def init_data(self):
-        self.x = np.random.uniform(-1, 1, [2, 4, 3, 5]).astype(np.float32)
-
-    def setUp(self):
-        self.dtype = np.uint16
-        self.init_data()
-        self.config()
-        self.out = self.op_forward(self.x)
-
-        self.inputs = {'X': convert_float_to_uint16(self.x)}
-        self.outputs = {'Out': self.out}
-        self.set_attrs()
-
-    def calculate_grads(self):
-        self.dx = self.op_grad(self.out, self.x)
-
-    def test_check_output(self):
-        self.check_output_with_place(core.CPUPlace())
-
-    def test_check_grad(self):
-        self.calculate_grads()
-        self.check_grad_with_place(
-            core.CPUPlace(), ["X"],
-            "Out",
-            user_defined_grads=[self.dx],
-            user_defined_grad_outputs=[convert_float_to_uint16(self.out)])
-
-
-class TestMKLDNNGeluErfBF16Op(TestMKLDNNSigmoidBF16Op):
-    def config(self):
-        self.op_type = "gelu"
-
-    def op_forward(self, x):
-        return gelu(x, False)
-
-    def op_grad(self, dout, x):
-        return (dout *
-                (0.5 + 0.5 * erf(x / np.sqrt(2)) +
-                 (x / np.sqrt(2 * np.pi) * np.exp(-0.5 * np.power(x, 2)))))
-
-
-class TestMKLDNNGeluErfDim2BF16Op(TestMKLDNNGeluErfBF16Op):
-    def init_data(self):
-        self.x = np.random.uniform(-1, 1, [11, 17]).astype(np.float32)
-
-
-class TestMKLDNNGeluTanhBF16Op(TestMKLDNNSigmoidBF16Op):
-    def config(self):
-        self.op_type = "gelu"
-
-    def op_forward(self, x):
-        return gelu(x, True)
-
-    def op_grad(self, dout, x):
-        grad_part = np.tanh(
-            np.sqrt(2 / np.pi) * (x + 0.044715 * np.power(x, 3)))
-        return dout * 0.5 * (1 + grad_part) * (1 + np.sqrt(2 / np.pi) *
-                                               (x + 0.134145 * np.power(x, 3)) *
-                                               (1 - grad_part))
-
-    def set_attrs(self):
-        self.attrs = {"use_mkldnn": True, "approximate": True}
-
-
-class TestMKLDNNGeluTanhDim2BF16Op(TestMKLDNNGeluTanhBF16Op):
-    def init_data(self):
-        self.x = np.random.uniform(-1, 1, [11, 17]).astype(np.float32)
-
-
 class TestMKLDNNTanhDim2(TestTanh):
     def setUp(self):
         super(TestMKLDNNTanhDim2, self).setUp()
